* develop: (339 commits)
  Prevent traps from casting periodic spells. This can be removed until proper event handling.
  Move flying npcs to the ground on dead. Close #216
  Fixed height and liquid loading. This was implemented by me in b2ab9a8a31209b15fcd8946aebd383297637da34 and 62724471c04a458036993acb6d5aaa823140feef (sorry for that).
  Added output for LandHeight to .gps command
  Header include vector in CRitual
  Header includes
  Corrected include Whooops.... here we go
  Header includes and forward declaration in CThreads and Database
  Header includes in Storage and CallBack
  Header includes for ObjectStorage
  Header includes for CObjectFactory and GameObject.h
  Include QuestMgr.h in GameObject.h
  Header includes
  Headerinclude IUpdatable to Object.h
  Added set vehicle faction to the faction of the first player passenger Close #205
  Assign m_difficulty to uint32 to prevent errors. Close #221
  Added GameObject classes:  *GameObject_Door  *GameObject_Button  *GameObject_QuestGiver  *GameObject_Chest  *GameObject_Trap  *GameObject_SpellFocus  *GameObject_Goober  *GameObject_FishingNode  *GameObject_Ritual  *GameObject_SpellCaster  *GameObject_FishingHole  *GameObject_Destructible
  Added structure gor gotype trap and spell_focus
  Prevent attack with invalid faction. Added because gos with type GAMEOBJECT_TYPE_SPELL_FOCUS  casting spells of linked gos with type GAMEOBJECT_TYPE_TRAP instead of spawning them.
  Implemented creature_proto_difficulty column money
  ...

Conflicts:
	.gitignore
	src/world/Chat.h
	src/world/DBC/DBCStores.cpp
	src/world/DBC/DBCStores.h
	src/world/GameObject.cpp
	src/world/GameObject.h
	src/world/Level3.cpp
	src/world/Master.cpp
	src/world/Object.h
	src/world/ObjectMgr.cpp
	src/world/ObjectMgr.h
	src/world/ScriptMgr.h
	src/world/TransporterHandler.cpp

+++ src/world/Faction.cpp
@@ -1,6 +1,6 @@
 /*
  * AscEmu Framework based on ArcEmu MMORPG Server
- * Copyright (C) 2014-2015 AscEmu Team <http://www.ascemu.org>
+ * Copyright (C) 2014-2016 AscEmu Team <http://www.ascemu.org>
  * Copyright (C) 2008-2012 ArcEmu Team <http://www.ArcEmu.org/>
  * Copyright (C) 2005-2007 Ascent Team
  *
@@ -55,6 +55,9 @@ bool isHostile(Object* objA, Object* objB)
     if (objB->IsUnit() && objB->HasFlag(UNIT_FIELD_FLAGS, 2 | 128 | 256 | 65536))
         return false;
 
+    if (!objB->m_faction || !objA->m_faction)
+        return false;
+
     uint32 faction = objB->m_faction->Mask;
     uint32 host = objA->m_faction->HostileMask;
 
@@ -68,7 +71,7 @@ bool isHostile(Object* objA, Object* objB)
         hostile = true;
 
     // check friend/enemy list
-    for (uint32 i = 0; i < 4; i++)
+    for (uint8 i = 0; i < 4; i++)
     {
         if (objA->m_faction->EnemyFactions[i] == objB->m_faction->Faction)
         {
@@ -174,6 +177,9 @@ bool isAttackable(Object* objA, Object* objB, bool CheckStealth)
             return true;
     }
 
+    if (objA->m_faction == nullptr)     // no faction, no kill (added because spell_caster gos should summon a trap instead of casting the spell directly.)
+        return false;
+
     if (objA->m_faction == objB->m_faction)    // same faction can't kill each other unless in ffa pvp/duel
         return false;
 
@@ -225,7 +231,7 @@ bool isCombatSupport(Object* objA, Object* objB)// B combat supports A?
         combatSupport = true;
     }
     // check friend/enemy list
-    for (uint32 i = 0; i < 4; i++)
+    for (uint8 i = 0; i < 4; i++)
     {
         if (objB->m_faction->EnemyFactions[i] == objA->m_faction->Faction)
         {
@@ -244,8 +250,8 @@ bool isCombatSupport(Object* objA, Object* objB)// B combat supports A?
 
 bool isAlliance(Object* objA)// A is alliance?
 {
-    FactionTemplateDBC* m_sw_faction = dbcFactionTemplate.LookupEntry(11);
-    FactionDBC* m_sw_factionDBC = dbcFaction.LookupEntry(72);
+    DBC::Structures::FactionTemplateEntry const* m_sw_faction = sFactionTemplateStore.LookupEntry(11);
+    DBC::Structures::FactionEntry const* m_sw_factionDBC = sFactionStore.LookupEntry(72);
     if (!objA)          // || objA->m_factionDBC == NULL || objA->m_faction == NULL
         return true;
 
@@ -260,7 +266,7 @@ bool isAlliance(Object* objA)// A is alliance?
         return false;
 
     // check friend/enemy list
-    for (uint32 i = 0; i < 4; i++)
+    for (uint8 i = 0; i < 4; i++)
     {
         if (objA->m_faction->EnemyFactions[i] == faction)
             return false;
@@ -273,7 +279,7 @@ bool isAlliance(Object* objA)// A is alliance?
         return false;
 
     // check friend/enemy list
-    for (uint32 i = 0; i < 4; i++)
+    for (uint8 i = 0; i < 4; i++)
     {
         if (objA->m_faction->EnemyFactions[i] == faction)
             return false;
